Add transfer object for webhook postback log entry
