+++ src/Events.Functions/Services/WebhookService.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Net;
 using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
@@ -7,7 +8,7 @@ using Altinn.Platform.Events.Functions.Extensions;
 using Altinn.Platform.Events.Functions.Models;
 using Altinn.Platform.Events.Functions.Models.Payloads;
 using Altinn.Platform.Events.Functions.Services.Interfaces;
-
+using CloudNative.CloudEvents;
 using Microsoft.Extensions.Logging;
 
 namespace Altinn.Platform.Events.Functions.Services
@@ -41,6 +42,7 @@ namespace Altinn.Platform.Events.Functions.Services
                 HttpResponseMessage response = await _client.PostAsync(envelope.Endpoint, httpContent);
 
                 // todo: post to storage controller with status code
+                await PostWebhookResponseForLogging(envelope.CloudEvent, response.StatusCode);
 
                 if (!response.IsSuccessStatusCode)
                 {
@@ -55,6 +57,17 @@ namespace Altinn.Platform.Events.Functions.Services
                 _logger.LogError(e, $"// Send to webhook with subscriptionId: {envelope.SubscriptionId} failed with error message {e.Message}");
                 throw;
             }
+        }
+
+        /// <summary>
+        /// Posts the provided cloud event to the logging endpoint
+        /// </summary>
+        /// <param name="cloudEvent">The cloud event associated with the logging entry <see cref="CloudEvent"/></param>
+        /// <param name="statusCode">The status code returned from the webhook endpoint</param>
+        /// <returns></returns>
+        internal async Task PostWebhookResponseForLogging(CloudEvent cloudEvent, HttpStatusCode statusCode)
+        {
+            await _client.PostAsync("/api/storage/events/logs", new StringContent(cloudEvent.Serialize(), Encoding.UTF8, "application/cloudevents+json"));
         }
 
         /// <summary>
